Add user reg, login, jwt token verification

+++ module/views.py
@@ -1,11 +1,47 @@
 from marshmallow import ValidationError
 from module import db
 from module import app
+from module import jwt
 from datetime import datetime
 from flask import make_response, request, jsonify
 import json
 from module.schemas import UserSchema, CategorySchema, RecordSchema, CurrencySchema
 from module.dbModels import User, Category, Record, Currency
+from passlib.hash import pbkdf2_sha256
+from flask_jwt_extended import JWTManager, jwt_required, create_access_token, get_jwt_identity
+
+
+#JWT Error Handlers
+
+@jwt.expired_token_loader
+def expired_token_callback(jwt_header, jwt_payload):
+   return (
+       jsonify({"message": "The token has expired.", "error": "token_expired"}),
+       401,
+   )
+
+@jwt.invalid_token_loader
+def invalid_token_callback(error):
+   return (
+       jsonify(
+           {"message": "Signature verification failed.", "error": "invalid_token"}
+       ),
+       401,
+   )
+
+@jwt.unauthorized_loader
+def missing_token_callback(error):
+   return (
+       jsonify(
+           {
+               "description": "Request does not contain an access token.",
+               "error": "authorization_required",
+           }
+       ),
+       401,
+   )
+
+
 
 # Flask routes
 
@@ -23,6 +59,7 @@ def healthcheck():
     return make_response((response, 200, {'Content-Type': 'application/json'}))
 
 @app.route('/currency', methods=['POST'])
+@jwt_required()
 def add_currency():
     currency_schema = CurrencySchema()
     try:
@@ -42,6 +79,7 @@ def get_currencies():
     return jsonify(currency_schema.dump(currencies)), 200
 
 @app.route('/currency/<int:currency_id>', methods=['DELETE'])
+@jwt_required()
 def delete_currency(currency_id):
     currency = Currency.query.get(currency_id)
     if currency:
@@ -50,8 +88,14 @@ def delete_currency(currency_id):
         return jsonify({}), 204
     return jsonify({'error': 'Currency not found'}), 404
 
+
 @app.route('/user/<int:user_id>/set_currency', methods=['POST'])
+@jwt_required()
 def set_user_currency(user_id):
+    current_user_id = get_jwt_identity()
+
+    if current_user_id != user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
     user = User.query.get(user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
@@ -78,34 +122,60 @@ def get_users():
 
 
 @app.route('/user', methods=['POST'])
-def add_user():
+def register():
     user_schema = UserSchema()
     try:
         user_data = user_schema.load(request.json)
+        user_data['password'] = pbkdf2_sha256.hash(user_data['password'])
+        new_user = User(**user_data)
+        db.session.add(new_user)
+        db.session.commit()
+        return jsonify({'user_id': new_user.id}), 201
     except ValidationError as err:
         return jsonify(err.messages), 400
 
-    new_user = User(name=user_data['name'])
-    db.session.add(new_user)
-    db.session.commit()
-    return jsonify({'user_id': new_user.id}), 201
+@app.route('/login', methods=['POST'])
+def login():
+    user_schema = UserSchema(only=["name", "password"])
+    try:
+        user_data = user_schema.load(request.json)
+        user = User.query.filter_by(name=user_data["name"]).first()
+        if user and pbkdf2_sha256.verify(user_data["password"], user.password):
+            access_token = create_access_token(identity=user.id)
+            return jsonify(access_token=access_token), 200
+        return jsonify({"error": "Invalid username or password"}), 401
+    except ValidationError as err:
+        return jsonify(err.messages), 400
 
 @app.route('/user/<int:user_id>', methods=['GET'])
+@jwt_required()
 def get_user(user_id):
+    current_user_id = get_jwt_identity()  # Get user ID from JWT token
+
+    if current_user_id != user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
     user = User.query.get(user_id)
     if user:
         user_schema = UserSchema()
         return jsonify(user_schema.dump(user)), 200
     return jsonify({'error': 'User not found'}), 404
 
+
+
 @app.route('/user/<int:user_id>', methods=['DELETE'])
+@jwt_required()
 def delete_user(user_id):
+    current_user_id = get_jwt_identity()
+
+    if current_user_id != user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
     user = User.query.get(user_id)
     if user:
         db.session.delete(user)
         db.session.commit()
         return jsonify({}), 204
     return jsonify({'error': 'User not found'}), 404
+
 
 @app.route('/category', methods=['GET'])
 def get_categories():
@@ -114,7 +184,9 @@ def get_categories():
     return jsonify(category_schema.dump(categories)), 200
 
 
+
 @app.route('/category', methods=['POST'])
+@jwt_required()
 def add_category():
     category_schema = CategorySchema()
     try:
@@ -127,7 +199,9 @@ def add_category():
     db.session.commit()
     return jsonify({'category_id': new_category.id}), 201
 
+
 @app.route('/category/<int:category_id>', methods=['DELETE'])
+@jwt_required()
 def delete_category(category_id):
     category = Category.query.get(category_id)
     if category:
@@ -136,7 +210,9 @@ def delete_category(category_id):
         return jsonify({}), 204
     return jsonify({'error': 'Category not found'}), 404
 
+
 @app.route('/record', methods=['POST'])
+@jwt_required()
 def add_record():
     record_schema = RecordSchema()
     try:
@@ -164,35 +240,62 @@ def add_record():
     db.session.commit()
     return jsonify({'record_id': new_record.id}), 201
 
+
 @app.route('/record', methods=['GET'])
+@jwt_required()
 def get_records():
-    user_id = request.args.get('user_id')
+    user_id_str = request.args.get('user_id')
+    current_user_id = get_jwt_identity()
+
+    try:
+        user_id = int(user_id_str) if user_id_str is not None else None
+    except ValueError:
+        return jsonify({'error': 'Invalid user ID'}), 400
+
+    # Compare the user IDs
+    if user_id is not None and current_user_id != user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
+
     category_id = request.args.get('category_id')
     records = Record.query.filter(
-        (Record.user_id == user_id) if user_id else True,
+        (Record.user_id == current_user_id),  # Use current_user_id for filtering
         (Record.category_id == category_id) if category_id else True
     ).all()
     record_schema = RecordSchema(many=True)
     return jsonify(record_schema.dump(records)), 200
 
+
 @app.route('/record/<int:record_id>', methods=['GET'])
+@jwt_required()
 def get_record_by_id(record_id):
+    current_user_id = get_jwt_identity()  # Get user ID from JWT token
     record = Record.query.get(record_id)
+
     if record is None:
         return jsonify({'message': 'Record not found'}), 404
 
+    if record.user_id != current_user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
 
     record_schema = RecordSchema()
     return jsonify(record_schema.dump(record)), 200
 
+
 @app.route('/record/<int:record_id>', methods=['DELETE'])
+@jwt_required()
 def delete_record(record_id):
+    current_user_id = get_jwt_identity()  # Get user ID from JWT token
     record = Record.query.get(record_id)
-    if record:
-        db.session.delete(record)
-        db.session.commit()
-        return jsonify({}), 204
-    return jsonify({'error': 'Record not found'}), 404
+
+    if record is None:
+        return jsonify({'error': 'Record not found'}), 404
+
+    if record.user_id != current_user_id:
+        return jsonify({'error': 'Unauthorized access'}), 403
+
+    db.session.delete(record)
+    db.session.commit()
+    return jsonify({}), 204
 
 
 if __name__ == '__main__':
